refactor : 토큰 재발급 api 요청 필요없는 로직 변경

+++ frontend/src/utils/api/token.js
@@ -7,7 +7,7 @@ const axiosConfig = axios.create({
     headers: {
         'Content-Type': 'application/json',
         Accept: 'application/json',
-        Authorization: `Bearer ${sessionStorage.getItem('accessToken')}`,
+        Authorization: `Bearer ${Cookies.get('accessToken')}`,
     },
 });
 
@@ -21,10 +21,6 @@ async function getAccessToken() {
             },
         });
 
-        // 토큰 유효성 검사할 때 필요
-        sessionStorage.setItem('accessToken', response.data.access);
-
-        // 토큰 유효기간 검사용으로 필요
         Cookies.set('accessToken', response.data.access, {
             path: '/',
             expires: expire, // 테스트 기준 5분 (하루 단위로 응답)
@@ -35,23 +31,6 @@ async function getAccessToken() {
             expires: 90, // 테스트 기준 90일
         });
 
-        return null;
-    } catch (error) {
-        return error.response;
-    }
-}
-
-// 액세스 토큰 존재 여부 확인
-async function isVerify() {
-    try {
-        // 액세스 토큰 존재 여부 확인
-        const response = await axiosUserConfig({
-            url: '/app/token/verify/',
-            data: {
-                token: sessionStorage.getItem('accessToken'),
-            },
-        });
-
         return response;
     } catch (error) {
         return error.response;
@@ -60,30 +39,25 @@ async function isVerify() {
 
 // access 토큰 유효성 검증 함수
 async function accessAvailableCheck() {
-    const isExistence = Cookies.get('accessToken');
+    const accessToken = Cookies.get('accessToken');
     const refreshToken = Cookies.get('refreshToken');
 
     try {
+        if (accessToken) return accessToken; // 액세스 토큰이 존재하면 그대로 반환
+
         // 리프레시 토큰이 존재하지 않으면 로그아웃
         if (!refreshToken) {
             sessionStorage.clear();
             window.open('/', '_self');
+        } else {
+            // 리프레시 토큰이 존재하고 액세스 토큰이 없다면 재발급 후 토큰 반환
+            const reAccessToken = await getAccessToken();
+            return reAccessToken;
         }
 
-        // 액세스 토큰 시간이 지났다면
-        // 유효성 검사 해본다
-        if (!isExistence) {
-            const response = await isVerify();
-
-            // 서버에서도 토큰이 존재하지 않는다면 재발급 요청
-            if (response.status === 401) {
-                await getAccessToken();
-            }
-        }
-
-        return sessionStorage.getItem('accessToken');
+        return null;
     } catch (error) {
-        return error;
+        return error.response;
     }
 }
 
@@ -91,9 +65,9 @@ async function accessAvailableCheck() {
 axiosConfig.interceptors.request.use(
     async config => {
         const axiosInstance = config;
-        const accessToken = await accessAvailableCheck();
+        await accessAvailableCheck();
 
-        axiosInstance.headers.Authorization = `Bearer ${accessToken}`;
+        axiosInstance.headers.Authorization = `Bearer ${Cookies.get('accessToken')}`;
         return axiosInstance;
     },
     error => {
